+++ docs/CHANGELOG.md
@@ -16,6 +16,7 @@ CURRENT
 - Fix RecordBrowser's field tooltip for select and multiselect fields
 - Fix Month View applet issue related to the daylight saving shift
 - Fix new langpack rule issue.
+- Remove duplicated codes from countries list and calling codes
 
 RELEASE 1.5.6-20140305
 -------